Refactor RepositoryPage.tsx
