+++ web/src/features/repositories/repository/RepositorySettingsInfo.tsx
@@ -1,0 +1,10 @@
+import {Stack} from "@mui/material";
+
+// TODO
+export function RepositorySettingsInfo() {
+    return (
+        <Stack>
+            Settings Info
+        </Stack>
+    )
+}
